Scroll to top of page when location changes from footer (#35)
This commit:

* Scrolls the page to the top when a footer link is clicked that changes
the location.

Why?

* Before this change the page would remain at the bottom when the
location changed.

Notes:

This change adds an `onClick` to the React Router `Link` component,
which already has a `to` prop. The `to` prop is required for `Link`
components, so I wasn't able to remove it and use the click handler for
changing location. I read [this
issue](https://github.com/ReactTraining/react-router/issues/5240) which
dives into a similar issue. What I think we want to be rendered here is
an anchor tag with an href value, which is what a `Link` with a `to`
prop gives us. Since the `onClick` is solely responsible for scrolling
to the top of the page I don't see it conflicting with the behavior of
the `to` prop.