Sticky TOC (#64)
* Sticky TOC

Position the TOC statically by default. When the window scrolls the header away, stick the TOC to the top of the page.

Had to drop the throttling down to every 10ms to eliminate some funky behavior in which the class doesn't change fast enough when user scrolls to top of the page.