Improve scroll performance on the schema page
Currently the app re-renders each OSQueryTable component while scrolling
causing a lot of slow-down on the page.

To resolve the issue we need to turn `OsQueryTable` into a pure
component so it's only rendered when its props change.

The second part of the issue is ensuring that its parent component
`Schema` always passes down the same table objects due to how JavaScript
equality works. The existing approach is to re-computer the filtered
tables on each render. This means each render `OsQueryTable` gets a
different instance of the same data which JavaScript considers not
equal. To fix this issue we now computer the filtered value once and
store it on state and only re-compute when necessary.