Prevent sidebar scroll chaining
This prevents the "scroll chaining" behavior when scrolling content
which scrolls outer content once the inner content can no longer be
scrolled. This was causing issues with the schema pages sidebar since
scroll to the top would scroll `body` and immediately update the sidebar
scroll.

This resolves the issue by setting `overflow: hidden;` on `body` when
the users mouse enters the sidebar and removes it when the mouse leaves.
Because body is unscrollable the scroll chaining "stops" at the sidebar.